move DEFAULT_MAX_SCALE to utils/crawler instead of types/crawler
