+++ frontend/src/pages/org/index.ts
@@ -6,11 +6,12 @@ import { ifDefined } from "lit/directives/if-defined.js";
 import type { ViewState } from "@/utils/APIRouter";
 import type { AuthState } from "@/utils/AuthService";
 import type { CurrentUser } from "@/types/user";
-import { DEFAULT_MAX_SCALE, type Crawl, type JobType } from "@/types/crawler";
+import type { Crawl, JobType } from "@/types/crawler";
 import type { OrgData } from "@/utils/orgs";
 import { isAdmin, isCrawler } from "@/utils/orgs";
 import LiteElement, { html } from "@/utils/LiteElement";
 import { needLogin } from "@/utils/auth";
+import { DEFAULT_MAX_SCALE } from "@/utils/crawler";
 import "./workflow-detail";
 import "./workflows-list";
 import "./workflows-new";
